apps/delulu_discord/pyproject.toml
[project]
name = "delulu-discord"
version = "0.1.0"
description = "Discord bot that dispatches Claude Code tasks to Modal sandboxes"
requires-python = ">=3.14"
dependencies = [
    "discord.py>=2.4,<3",
    "modal>=1.0",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
delulu-discord = "delulu_discord.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/delulu_discord"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
